refactor: Change the directory structure
