+++ application/millage-client/src/pages/App.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
-import Header from './header/Header';
-import Footer from './footer/Footer';
-import Login from './login/Login';
+import Header from '@components/Header';
+import Footer from '@components/Footer';
+import Login from '@pages/Login';
 import './App.css';
 
 function App() {
